supporting_scripts/generate_code_cov_table/requirements.txt
requests==2.27.1
gitpython==3.1.31
beautifulsoup4==4.12.2
pyperclip==1.8.2
python-dotenv==1.0.0